+++ node/resolvers/fieldResolvers.ts
@@ -6,16 +6,6 @@ import {
   COST_CENTER_FIELDS,
 } from '../mdSchema'
 
-const QUERIES = {
-  getRole: `query role($id: ID!) {
-    getRole(id: $id) @context(provider: "vtex.storefront-permissions") {
-        id
-        name
-        slug
-    }
-  }`,
-}
-
 export const organizationName = async (
   { orgId }: { orgId: string },
   _: any,
@@ -88,21 +78,12 @@ export const role = async (
   ctx: Context
 ) => {
   const {
-    clients: { graphQLServer },
+    clients: { storefrontPermissions },
     vtex: { logger },
   } = ctx
 
-  const roleData = await graphQLServer
-    .query(
-      QUERIES.getRole,
-      { id: roleId },
-      {
-        persistedQuery: {
-          provider: 'vtex.storefront-permissions@1.x',
-          sender: 'vtex.b2b-organizations@0.x',
-        },
-      }
-    )
+  const roleData = await storefrontPermissions
+    .getRole(roleId)
     .then((result: any) => {
       return result.data.getRole
     })