+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [0.11.0] - 2022-03-24
+
+### Fixed
+
+- In `impersonateUser` resolver, if `clId` is provided but `userId` is not, check CL document to see if userId exists there and then update storefront-permissions' user record
+- Similarly, when editing a user, if `userId` is null, check CL document to see if userId exists there and then sync it with storefront-permissions
+
+### Changed
+
+- Use storefront-permissions' GraphQL interface directly instead of sending requests through `vtex.graphql-server`
+
 ## [0.10.0] - 2022-03-08
 
 ### Added